invoked developer agent to fix data layer findings

+++ app/src/main/java/com/delve/hungrywalrus/data/remote/usda/UsdaDto.kt
@@ -9,7 +9,7 @@ data class UsdaSearchResponse(
 
 @Serializable
 data class UsdaFood(
-    val fdcId: Int,
+    val fdcId: Long,
     val description: String,
     val foodNutrients: List<UsdaNutrient> = emptyList(),
 )
